Clang-tidy

+++ include/Cellnta/Algorithms/AlgoBase.h
@@ -18,16 +18,16 @@ class AlgoBase {
 
   virtual void Update() = 0;
   virtual void LoadWorld(RenderData* data) = 0;
-  virtual void SetDimension(size_t dim) = 0;
+  virtual void SetDimension(int dim) = 0;
 
   virtual void SetCell(const Cell& cell) = 0;
   virtual void SetCell(const std::vector<Cell>& cells) = 0;
   virtual state_t GetCell(const Cell::Vec& pos) = 0;
 
-  size_t GetDimensions() { return p_dim; }
+  int GetDimensions() { return p_dim; }
 
-  void SetStep(size_t step) { m_step = step; }
-  size_t GetStep() const { return m_step; }
+  void SetStep(int step) { m_step = step; }
+  int GetStep() const { return m_step; }
 
   AlgoType GetType() const { return m_type; }
   bool NeedLoadWorld() const { return p_needLoadInRenderer; }
@@ -42,10 +42,10 @@ class AlgoBase {
 
  protected:
   bool p_needLoadInRenderer = false;
-  size_t p_dim = 0;
+  int p_dim = 0;
 
  private:
-  size_t m_step = 1;
+  int m_step = 1;
   const AlgoType m_type;
 };
