+++ include/Cellnta/Renderer/Transform.h
@@ -7,16 +7,16 @@
 
 namespace Cellnta {
 
-Eigen::MatrixXf LookAt(const size_t N, const Eigen::VectorXf& from,
+Eigen::MatrixXf LookAt(const int N, const Eigen::VectorXf& from,
                        const Eigen::VectorXf& to, const Eigen::MatrixXf& ups);
-Eigen::MatrixXf Perspective(const size_t N, double eyeAngle, double nearPlane,
+Eigen::MatrixXf Perspective(const int N, double eyeAngle, double nearPlane,
                             double farPlane, double aspect);
 
-void NProject(NCellStorage& cells, const size_t cameraDim,
+void NProject(NCellStorage& cells, const int cameraDim,
               const Eigen::MatrixXf& viewProj, bool perspective);
-void NProject(HypercubeStorage& cube, const size_t cameraDim,
+void NProject(HypercubeStorage& cube, const int cameraDim,
               const Eigen::MatrixXf& viewProj, bool perspective);
 
-Eigen::MatrixXf NRotate(size_t N, size_t axis1, size_t axis2, float angle);
+Eigen::MatrixXf NRotate(int N, int axis1, int axis2, float angle);
 
 }  // namespace Cellnta
